sql: move a test from rowexec to execinfra
To match the new location of the code under test.

Release justification: test only

Release note: None
